Refactor layouts to isolate login page
